All components put in separate folders as well as updated imports are now implemented.
