+++ src/components/AppLayout/AppLayout.js
@@ -1,9 +1,9 @@
 'use client';
-import Header from './Header';
-import NavigationModal from './NavigationModal/NavigationModal';
-import Footer from './Footer';
+import Header from '../Header/Header';
+import NavigationModal from '../NavigationModal/NavigationModal';
+import Footer from '../Footer/Footer';
 import { useState } from 'react';
-import AuthContextProvider from '../context/AuthContext';
+import AuthContextProvider from '../../context/AuthContext';
 
 export default function AppLayout({ children }) {
   const [isModalOpen, setIsModalOpen] = useState(false);
