+++ src/components/icons/ArrowRightIcon.js
